configs: Build simple DeviceConfig provisioning service
We need a simple service to provision DeviceConfig for compatibility
with the Device Personalization Services app on Pixels, and we can't let
Google Play Services do the provisioning because it overrides some of
our other default settings. Build our custom service to do the
provisioning from resource overlay values instead.